+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Version 0.5.0 (Planned)
 
 - If there are multiple pots (main + side), ``pokerkit.state.State.push_chips``
   must be called multiple times.
+- Custom automations is passed through constructor for ``pokerkit.notation.HandHistory''.
 
 Version 0.4.6 (January 8, 2024)
 -------------------------------